Remove EffectsComposer post processing pipeline.
Resolves initial problem with https://github.com/ghemingway/cad.js/issues/16

The post processing pipeline was added initially, many moons ago,
to allow the Screen Space Ambient Occlusion (SSAO) effect. This
effect has since proved problematic and been removed. The post
processing pipeline still should allow for Antialiasing on platforms
that do not natively support it but the current pipeline was broken.

This commit removes the pipeline so that systems use their native
anti-aliasing if available and degrade gracefully to no anti-aliasing,
like the older iPhone I'm testing on.

This does not resolve the flickering/occluded UI elements issue
when rotating. That issue is something else to be picked up next.